pyproject.toml
[tool.poetry]
name = "terminus-web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.3"
django-rest-framework = "^0.1.0"
channels = "^3.0.3"
uvloop = "^0.15.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
